+++ BMExtendablePageController/BMExtendablePageController.m
@@ -131,7 +131,7 @@
     _temporaryDisabled = TRUE;
     
     NSAssert(_selectedIndex < _pages.count,
-             @"_selectedIndex is %i, but there are only %i pages",_selectedIndex,_pages.count);
+             @"_selectedIndex is %i, but there are only %i pages",_selectedIndex,(int)_pages.count);
     
     VIEW* currentView = [(VIEW_CONTROLLER*)[_pages objectAtIndex:_selectedIndex] view];
